refactor: move backup commands to one spot
